Meal Seating Fix (#1524)
* Meal Seating Fix
This issue is for one of the on-boarding trainees. We need to get the meal order print out in alphabetical order down the column instead of being alphabetical by row. If you look at the meal seating chart in the cafeteria you will immediately see what I am talking about.

 Recently, when the meal seating sisters generated based on date of birth it did it by year instead of month/day. Obviously, this is awkward as all the 30 years olds were sitting together. Needs to be changed

 Bonus - - Add more variables for them to generate meal seating by - added sort by email, TA

* Fixed a bug where first names would be sorted randomly. Now properly sorts by last_name,first_name.